Integrate Leigh's changes and change csv generator class to be a report class
