+++ app/jobs/school_group_meter_report_job.rb
@@ -6,7 +6,7 @@ class SchoolGroupMeterReportJob < ApplicationJob
   end
 
   def perform(to:, school_group:, meter_scope: {})
-    csv = SchoolGroups::Meters::CsvGenerator.new(school_group, meter_scope)
-    AdminMailer.with(to: to, school_group: school_group, filename: csv.filename, csv: csv.content).school_group_meters_report.deliver
+    meter_report = SchoolGroups::MeterReport.new(school_group, meter_scope)
+    AdminMailer.with(to: to, school_group: school_group, filename: meter_report.csv_filename, csv: meter_report.csv).school_group_meters_report.deliver
   end
 end
